requirements.txt
aioredis==1.3.1
uvicorn==0.15.0
elasticsearch-async==6.2.0
fastapi-pagination==0.8.3
fastapi==0.68.1
orjson==3.6.3
pydantic[dotenv]